Polished

+++ bel/temp/folder_cleanup.py
@@ -12,11 +12,7 @@ for r, d, f in os.walk(res_dir, topdown=False):
 
 for r, d, f in os.walk(res_dir, topdown=False):
     # Adds the data files to the lists, which will be loaded later
-    if 'bkt.npy' in f and 'pz.npy' in f and 'hk0.npy' in f:
-        pass
-    else:  # If one of the files is missing, deletes the sub-folder
-        try:
-            if r != res_dir:  # Make sure to not delete the main results directory !
-                shutil.rmtree(r)
-        except TypeError:
-            pass
+    if 'mt3d_link.ftl' in f:
+        if r != res_dir:  # Make sure to not delete the main results directory !
+            print('removed 1 folder')
+            shutil.rmtree(r)
